This is first page scraping of GitHub
It is rough of scrapping of first page. And store the data in csv file and store the data in data folder and all data like url, repository name and title. Its a use BeautifulSoup and request library.
I am able to scrape the data but it is my first project. I will provide the new jupyter notebook those are presented in a well manner and this is hit and trial jupyter notebook.